+++ app/src/main/java/com/artemissoftware/tasky/agenda/data/remote/dto/EventAttendeeDto.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.tasky.data.remote.dto.agenda
+package com.artemissoftware.tasky.agenda.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
