+++ app/src/main/java/com/artemissoftware/tasky/authentication/data/remote/dto/LoginResponseDto.kt
@@ -1,13 +1,13 @@
-package com.artemissoftware.tasky.data.remote.dto.agenda
+package com.artemissoftware.tasky.authentication.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
 
-data class AttendeeProfileDto(
-    @SerializedName("email")
-    val email: String,
+data class LoginResponseDto(
     @SerializedName("fullName")
     val fullName: String,
+    @SerializedName("token")
+    val token: String,
     @SerializedName("userId")
     val userId: String
 )
