+++ core/src/main/java/com/artemissoftware/core/data/remote/exceptions/TaskyNetworkException.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.tasky.data.remote.exceptions
+package com.artemissoftware.core.data.remote.exceptions
 
 data class TaskyNetworkException (
     val code: Int = NetworkErrors.GENERIC_API_ERROR.first,
